+++ lib/main/tree/main-diameter.cpp
@@ -15,7 +15,7 @@ void action_init() {
 
 void action_edge() {
     size_t u, v;
-    unsigned long long w;
+    uint64_t w;
     cin >> u >> v >> w;
     if (u >= solver.size() ||
         v >= solver.size())
@@ -29,7 +29,7 @@ void action_edge() {
 
 void action_solve() {
     size_t u, v;
-    unsigned long long w;
+    uint64_t w;
     tie(u, v, w) = solver.solve();
     cout << u << " " << v << " " << w << endl;
 }
